Split the old and new gravity headers
