+++ libgadget/run.c
@@ -251,8 +251,6 @@ begrun(const int RestartSnapNum, struct header_data * head)
 
     init_cooling_and_star_formation(All.CoolingOn, All.StarformationOn, &All.CP, head->MassTable[0], head->BoxSize, units);
 
-    gravshort_fill_ntab(All.ShortRangeForceWindowType, All.Asmth);
-
     if(All.LightconeOn)
         lightcone_init(&All.CP, head->TimeSnapshot, head->UnitLength_in_cm, All.OutputDir);
